Moved tests for minimize and resolution to respective files
